requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
arrowhead-client==0.4.4a0
async-timeout==4.0.2
attrs==21.4.0
blessings==1.7
bpython==0.22.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cmd2==2.4.1
cryptography==37.0.4
curtsies==0.3.10
cwcwidth==0.1.6
cycler==0.11.0
deepdiff==5.8.1
fastapi==0.79.0
Flask==2.1.3
fonttools==4.34.4
frozenlist==1.3.0
greenlet==1.1.2
grpcio==1.47.0
grpcio-tools==1.47.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
jwcrypto==0.8
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
numpy==1.23.1
ordered-set==4.1.0
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pykka==3.1.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.28
PyYAML==6.0
requests==2.28.1
shiboken6==6.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uvicorn==0.18.2
wcwidth==0.2.5
Werkzeug==2.1.2
yarl==1.7.2
